+++ tests/test_datastructures.py
@@ -37,6 +37,7 @@ from werkzeug._compat import iterlistvalues
 from werkzeug._compat import itervalues
 from werkzeug._compat import PY2
 from werkzeug._compat import text_type
+from werkzeug.datastructures import LanguageAccept
 from werkzeug.datastructures import Range
 from werkzeug.exceptions import BadRequestKeyError
 
@@ -911,12 +912,12 @@ def make_call_asserter(func=None):
 
     :param func: Additional callback for each function call.
 
-    >>> assert_calls, func = make_call_asserter()
-    >>> with assert_calls(2):
-    ...    func()
-    ...    func()
+    .. code-block:: python
+        assert_calls, func = make_call_asserter()
+        with assert_calls(2):
+            func()
+            func()
     """
-
     calls = [0]
 
     @contextmanager
@@ -1098,6 +1099,30 @@ class TestMIMEAccept(object):
         accept = self.storage_class([("*/*", 1), ("text/html", 1), ("image/*", 1)])
         assert accept.best_match(["image/png", "text/html"]) == "text/html"
         assert accept.best_match(["text/plain", "image/png"]) == "image/png"
+
+
+class TestLanguageAccept(object):
+    @pytest.mark.parametrize(
+        ("values", "matches", "default", "expect"),
+        (
+            ([("en-us", 1)], ["en"], None, "en"),
+            ([("en", 1)], ["en_US"], None, "en_US"),
+            ([("en-GB", 1)], ["en-US"], None, None),
+            ([("de_AT", 1), ("de", 0.9)], ["en"], None, None),
+            ([("de_AT", 1), ("de", 0.9), ("en-US", 0.8)], ["de", "en"], None, "de"),
+            ([("de_AT", 0.9), ("en-US", 1)], ["en"], None, "en"),
+            ([("en-us", 1)], ["en-us"], None, "en-us"),
+            ([("en-us", 1)], ["en-us", "en"], None, "en-us"),
+            ([("en-GB", 1)], ["en-US", "en"], "en-US", "en"),
+            ([("de_AT", 1)], ["en-US", "en"], "en-US", "en-US"),
+            ([("aus-EN", 1)], ["aus"], None, "aus"),
+            ([("aus", 1)], ["aus-EN"], None, "aus-EN"),
+        ),
+    )
+    def test_best_match_fallback(self, values, matches, default, expect):
+        accept = LanguageAccept(values)
+        best = accept.best_match(matches, default=default)
+        assert best == expect
 
 
 class TestFileStorage(object):